selene_multiple_browsers/pyproject.toml
[tool.poetry]
name = "selene-multiple-browsers"
version = "0.1.0"
description = ""
authors = ["dskard"]
license = "MIT"
readme = "README.md"
packages = [{include = "selene_multiple_browsers", from = "src"},]

[tool.poetry.dependencies]
python = "^3.10"
pytest-selenium = {git = "https://github.com/pytest-dev/pytest-selenium.git", rev = "47f4ef3"}
selene = {git = "https://github.com/yashaka/selene.git", rev = "2.0.0rc1"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.0"
pdbpp = "^0.10.3"
python-semantic-release = "^7.33.2"
shownode = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
version_variable = [
    "src/selene_multiple_browsers/__init__.py:__version__"
]
version_toml = [
    "pyproject.toml:tool.poetry.version"
]
major_on_zero = false
branch = "main"
upload_to_PyPI = true
upload_to_release = true
build_command = "pip install poetry && poetry build"
